+++ werkzeug/contrib/fixers.py
@@ -137,11 +137,13 @@ class ProxyFix(object):
         forwarded_for = getter('HTTP_X_FORWARDED_FOR', '').split(',')
         forwarded_host = getter('HTTP_X_FORWARDED_HOST', '')
         forwarded_port = getter('HTTP_X_FORWARDED_PORT', '')
+        forwarded_prefix = getter('HTTP_X_FORWARDED_PREFIX', '')
         environ.update({
             'werkzeug.proxy_fix.orig_wsgi_url_scheme': getter('wsgi.url_scheme'),
             'werkzeug.proxy_fix.orig_remote_addr': getter('REMOTE_ADDR'),
             'werkzeug.proxy_fix.orig_http_host': getter('HTTP_HOST'),
             'werkzeug.proxy_fix.orig_server_port': getter('SERVER_PORT'),
+            'werkzeug.proxy_fix.orig_script_name': getter('SCRIPT_NAME'),
         })
         forwarded_for = [x for x in [x.strip() for x in forwarded_for] if x]
         remote_addr = self.get_remote_addr(forwarded_for)
@@ -160,6 +162,8 @@ class ProxyFix(object):
                 environ['SERVER_PORT'] = forwarded_port
         if forwarded_proto:
             environ['wsgi.url_scheme'] = forwarded_proto
+        if forwarded_prefix:
+            environ['SCRIPT_NAME'] = forwarded_prefix
         return self.app(environ, start_response)